[FIX] BTK : One cl warning.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@230 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/VisSupport/VTK/btkVTKMarkersFramesSource.cpp
@@ -566,7 +566,7 @@ namespace btk
         }
         points->Delete();
         centers->Delete();
-        for (int i = 0 ; i < this->mp_TrajectoryIds->size() ; ++i)
+        for (int i = 0 ; i < static_cast<int>(this->mp_TrajectoryIds->size()) ; ++i)
         {
           this->mp_TrajectoryIds->operator[](i)->InsertNextCell(trajectoryPaths[i]);
           trajectoryPaths[i]->Delete();
